requirements.txt
mcp>=1.20.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
streamlit>=1.32.2
google-genai>=0.2.0
chromadb>=0.5.0
pypdf>=4.0.1
python-docx>=0.8.11
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.27.0
sentence-transformers>=2.2.2
